move some things out of util
