+++ bgcArgoDMQC/util/stats.py
@@ -1,0 +1,89 @@
+
+import sys
+import numpy as np
+
+def aic(data, resid):
+    '''
+    Function to calculate the Akiake Information Criteria (AIC) as a metric
+    for assessing the appropriate number of breakpoints in the calculation of
+    drifts in O2 gains.
+    
+    Args:
+    
+    Returns:
+    
+    Author:   
+        Christopher Gordon
+        Fisheries and Oceans Canada
+        chris.gordon@dfo-mpo.gc.ca
+    
+    Acknowledgement: this code is adapted from the SOCCOM SAGE_O2Argo matlab
+    code, available via https://github.com/SOCCOM-BGCArgo/ARGO_PROCESSING,
+    written by Tanya Maurer & Josh Plant
+    
+    Last update: 2020-10-27
+    
+    Change log:
+        - 2020-10-27: fixed print output
+    '''
+
+    # calculate AIC
+    SSE = np.sum(resid**2) # sum square errors
+    n = resid.shape[0]
+    m = data.shape[0] - 1 # do not include first cycle
+    K = 2*m + 2
+
+    # valid data parameters? see Jones & Day (1995)
+    is_valid = n/4 - 1
+    if m > is_valid:
+        aic_value = np.nan
+        sys.stdout.write('n >> K, cannot caclculate AIC, setting AIC = NaN\n')
+    else:
+        # formula ref. Jones & Day (1995), Owens & Wong (2009)
+        aic_value = np.log(SSE/n) + (n+K)/(n-K-2)
+
+    return aic_value
+
+
+def bic(data, resid):
+    '''
+    Function to calculate the Bayesian Information Criteria (BIC) as a metric
+    for assessing the appropriate number of breakpoints in the calculation of
+    drifts in O2 gains.
+    
+    Args:
+    
+    Returns:
+    
+    Author:   
+        Christopher Gordon
+        Fisheries and Oceans Canada
+        chris.gordon@dfo-mpo.gc.ca
+    
+    Acknowledgement: this code is adapted from the SOCCOM SAGE_O2Argo matlab
+    code, available via https://github.com/SOCCOM-BGCArgo/ARGO_PROCESSING,
+    written by Tanya Maurer & Josh Plant
+    
+    Last update: 2020-10-27
+    
+    Change log:
+        - 2020-10-27: fixed print output
+
+    '''
+
+    # calculate BIC
+    errorlim = 0 # cap on residuals, useful for noisy pH and nitrate data
+    SSE = np.sum(resid**2) # sum square errors
+    n = resid.shape[0]
+    m = data.shape[0] - 1 # do not include first cycle
+    K = 2*m + 2
+
+    # valid data parameters? see Jones & Day (1995)
+    is_valid = n/4 - 1
+    if m > is_valid:
+        bic_value = np.nan
+        sys.stdout.write('n >> K, cannot caclculate BIC, setting BIC = NaN\n')
+    else:
+        bic_value = np.log(1/(n*SSE) + errorlim**2) + K*np.log(n)/n
+
+    return bic_value
